requirements.txt
asgiref==3.5.2
Django==4.1.3
django-widget-tweaks==1.4.12
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
